+++ derive/src/merged_subscription.rs
@@ -73,6 +73,7 @@ pub fn generate(object_args: &args::MergedSubscription) -> GeneratorResult<Token
                         keys: ::std::option::Option::None,
                         visible: #visible,
                         is_subscription: true,
+                        rust_typename: ::std::any::type_name::<Self>(),
                     }
                 })
             }